+++ client/src/views/PeerReview/PeerReview.js
@@ -115,7 +115,7 @@ const PeerReview = props => {
         autoDismiss: true
       })
     }
-  }, [peerReviewDeleteSuccess, peerReviewDeleteError])
+  }, [peerReviewDeleteSuccess, peerReviewDeleteError, addToast])
 
   const viewDetailHandler = details => {
     setReviewDetails(details)
@@ -133,12 +133,12 @@ const PeerReview = props => {
     )
     filteredEmployee.map((review, key) => {
       tempArr.push([
-        <a
+        <span
           className={classes.showPointer}
           onClick={() => viewDetailHandler(review)}
         >
           {review.employee_under_review}
-        </a>,
+        </span>,
         review.employee_reviewing,
         review.project,
         review.to_date.slice(0, 10),
@@ -180,65 +180,65 @@ const PeerReview = props => {
           ClickHandler={detailsSwitchHandler}
         ></CreatePeerForm>
       ) : (
-        <GridContainer>
-          <Grid xs={1} sm={1} md={1} className={classes.grid} item>
-            <InputLabel>Search By:</InputLabel>
-          </Grid>
-          <GridItem xs={5} sm={5} md={5}>
-            <FormControl className={classes.formControl}>
-              <InputLabel htmlFor="SelectEmployee"> Select Employee</InputLabel>
-              <Select
-                value={selectedEmployee}
-                onChange={changeHandler}
-                inputProps={{
-                  name: 'SelectEmployee',
-                  id: 'SelectEmployee'
-                }}
-              >
-                <MenuItem className={classes.hoverEffect} value="">
-                  <em>None</em>
-                </MenuItem>
-                {employeeData
-                  ? employeeData.map((prop, key) => {
-                    return prop.status !== 'Inactive' ? ( 
-                     <MenuItem
-                        className={classes.hoverEffect}
-                        value={prop.userName}
-                        key={key}
-                      >
-                        {prop.userName}
-                      </MenuItem>
-                    ) : null
+          <GridContainer>
+            <Grid xs={1} sm={1} md={1} className={classes.grid} item>
+              <InputLabel>Search By:</InputLabel>
+            </Grid>
+            <GridItem xs={5} sm={5} md={5}>
+              <FormControl className={classes.formControl}>
+                <InputLabel htmlFor="SelectEmployee"> Select Employee</InputLabel>
+                <Select
+                  value={selectedEmployee}
+                  onChange={changeHandler}
+                  inputProps={{
+                    name: 'SelectEmployee',
+                    id: 'SelectEmployee'
+                  }}
+                >
+                  <MenuItem className={classes.hoverEffect} value="">
+                    <em>None</em>
+                  </MenuItem>
+                  {employeeData
+                    ? employeeData.map((prop, key) => {
+                      return prop.status !== 'Inactive' ? (
+                        <MenuItem
+                          className={classes.hoverEffect}
+                          value={prop.userName}
+                          key={key}
+                        >
+                          {prop.userName}
+                        </MenuItem>
+                      ) : null
                     })
-                  : null}
-              </Select>
-            </FormControl>
-          </GridItem>
-          <GridItem style={{ textAlign: 'end' }} xs={6} sm={6} md={6}>
-            <Button color="primary" onClick={createPeerHandler}>
-              Create Peer
+                    : null}
+                </Select>
+              </FormControl>
+            </GridItem>
+            <GridItem style={{ textAlign: 'end' }} xs={6} sm={6} md={6}>
+              <Button color="primary" onClick={createPeerHandler}>
+                Create Peer
             </Button>
-          </GridItem>
-          <GridItem xs={12} sm={12} md={12}>
-            <Card plain>
-              <CardHeader plain color="primary">
-                <h4 className={classes.cardTitleWhite}>PEER REVIEW</h4>
-              </CardHeader>
-              <CardBody>
-                <Table
-                  tableHeaderColor="gray"
-                  tableHead={peerReviewListingHeader}
-                  tableData={tempArr || null}
-                  addLinks={links}
-                  updateUser={updateUser}
-                  deleteUser={deleteUser}
-                  showLink={false}
-                />
-              </CardBody>
-            </Card>
-          </GridItem>
-        </GridContainer>
-      )}
+            </GridItem>
+            <GridItem xs={12} sm={12} md={12}>
+              <Card plain>
+                <CardHeader plain color="primary">
+                  <h4 className={classes.cardTitleWhite}>PEER REVIEW</h4>
+                </CardHeader>
+                <CardBody>
+                  <Table
+                    tableHeaderColor="gray"
+                    tableHead={peerReviewListingHeader}
+                    tableData={tempArr || null}
+                    addLinks={links}
+                    updateUser={updateUser}
+                    deleteUser={deleteUser}
+                    showLink={false}
+                  />
+                </CardBody>
+              </Card>
+            </GridItem>
+          </GridContainer>
+        )}
       <Dialog title="Delete Peer" modal={true} open={showDelDialog}>
         <DialogActions>
           <GridItem xs={12} sm={12} md={12}>
